+++ CHANGELOG.md
@@ -14,6 +14,18 @@
 5. **Simpletools\Db\Cassandra\Batch**
     1. Added `->constraint()` to force constraint index
 
+### 0.3.28 (2020-02-11)
+1. **Simpletools\Db\Cassandra\TempTable**
+    1. Added auto create
+    
+### 0.3.27 (2020-02-11)
+1. **Simpletools\Db\Cassandra\Schema**
+    1. Added support for keyspace on definition, defaults to client default keyspace
+    
+### 0.3.26 (2020-02-11)
+1. **Simpletools\Db\Cassandra\TempTable**
+    1. Introduced new TempTable() functionality
+    
 ### 0.3.24 (2020-02-04)
 1. **Simpletools\Db\Cassandra\Query**
     1. Added ->increase() and ->decrease() for counter table